+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Plugin 🕹️
 - Change icons_preferences, improve build time
 - Fix logger TypeError after enabling from ignored_packages, avoiding Sublime Text restart
+- Fix listener and add_on_change not starting after enabling via '.sublime-package' during upgrade
+
 
 ## [0.4.5] - 2025-01-22